Add warning flags to Makefile (#48)

+++ project/src/common/game.hpp
@@ -2,26 +2,27 @@
 #define GAME_H
 
 #include <cstddef>
+#include <cstdint>
 
 class Game {
  protected:
-  int numErrors = 0;
-  int currentTrial = 1;
+  uint32_t numErrors = 0;
+  uint32_t currentTrial = 1;
   bool isActive = true;
-  int playerId;
-  int wordLen;
-  int maxErrors;
+  uint32_t playerId;
+  uint32_t wordLen;
+  uint32_t maxErrors;
   char *wordProgress;
 
  public:
-  int getPlayerId();
-  size_t getWordLen();
-  int getMaxErrors();
-  int getCurrentTrial();
+  uint32_t getPlayerId();
+  uint32_t getWordLen();
+  uint32_t getMaxErrors();
+  uint32_t getCurrentTrial();
   char *getWordProgress();
-  int getNumErrors();
-  void updateWordChar(int index, char letter);
-  void updateCurrentTrial(int num);
+  uint32_t getNumErrors();
+  void updateWordChar(uint32_t index, char letter);
+  void updateCurrentTrial(uint32_t num);
   void updateNumErrors();
   bool getIsActive();
 };
